+++ tests/machine.test.ts
@@ -2,6 +2,7 @@ import {
     BackupDecryptionKey,
     CrossSigningStatus,
     DecryptedRoomEvent,
+    DecryptedToDeviceEvent,
     DecryptionErrorCode,
     DecryptionSettings,
     DeviceId,
@@ -12,6 +13,7 @@ import {
     EventId,
     getVersions,
     InboundGroupSession,
+    InvalidToDeviceEvent,
     KeysBackupRequest,
     KeysClaimRequest,
     KeysQueryRequest,
@@ -19,7 +21,10 @@ import {
     MaybeSignature,
     MegolmDecryptionError,
     OlmMachine,
+    OtherUserIdentity,
     OwnUserIdentity,
+    PlainTextToDeviceEvent,
+    ProcessedToDeviceEventType,
     RequestType,
     RoomId,
     RoomKeyWithheldInfo,
@@ -33,7 +38,7 @@ import {
     ToDeviceRequest,
     TrustRequirement,
     UserId,
-    OtherUserIdentity,
+    UTDToDeviceEvent,
     VerificationRequest,
     Versions,
 } from "@matrix-org/matrix-sdk-crypto-wasm";
@@ -306,10 +311,12 @@ describe(OlmMachine.name, () => {
         const oneTimeKeyCounts = new Map();
         const unusedFallbackKeys = new Set();
 
-        const receiveSyncChanges = JSON.parse(
-            await m.receiveSyncChanges(toDeviceEvents, changedDevices, oneTimeKeyCounts, unusedFallbackKeys),
+        const receiveSyncChanges = await m.receiveSyncChanges(
+            toDeviceEvents,
+            changedDevices,
+            oneTimeKeyCounts,
+            unusedFallbackKeys,
         );
-
         expect(receiveSyncChanges).toEqual([]);
     });
 
@@ -319,10 +326,12 @@ describe(OlmMachine.name, () => {
         const changedDevices = new DeviceLists();
         const oneTimeKeyCounts = new Map();
 
-        const receiveSyncChanges = JSON.parse(
-            await m.receiveSyncChanges(toDeviceEvents, changedDevices, oneTimeKeyCounts, undefined),
+        const receiveSyncChanges = await m.receiveSyncChanges(
+            toDeviceEvents,
+            changedDevices,
+            oneTimeKeyCounts,
+            undefined,
         );
-
         expect(receiveSyncChanges).toEqual([]);
     });
 
@@ -333,10 +342,12 @@ describe(OlmMachine.name, () => {
         const oneTimeKeyCounts = new Map();
         const unusedFallbackKeys = new Set();
 
-        const receiveSyncChanges = JSON.parse(
-            await m.receiveSyncChanges(toDeviceEvents, changedDevices, oneTimeKeyCounts, unusedFallbackKeys),
+        const receiveSyncChanges = await m.receiveSyncChanges(
+            toDeviceEvents,
+            changedDevices,
+            oneTimeKeyCounts,
+            unusedFallbackKeys,
         );
-
         expect(receiveSyncChanges).toEqual([]);
 
         const outgoingRequests = await m.outgoingRequests();
@@ -1622,18 +1633,17 @@ describe(OlmMachine.name, () => {
 
             let toDeviceEvents = [someClearEvent];
 
-            const received = await m.receiveSyncChanges(
+            const receivedToDeviceArray = await m.receiveSyncChanges(
                 JSON.stringify(toDeviceEvents),
                 new DeviceLists(),
                 new Map<string, number>(),
                 undefined,
             );
-
-            const receivedToDevices = JSON.parse(received);
-            expect(receivedToDevices).toBeInstanceOf(Array);
-            const receivedToDeviceArray: Array<any> = receivedToDevices;
             expect(receivedToDeviceArray.length).toBe(1);
-            const toDeviceEvent = receivedToDeviceArray[0];
+            const processed = receivedToDeviceArray[0];
+            expect(processed.type).toEqual(ProcessedToDeviceEventType.PlainText);
+            expect(processed).toBeInstanceOf(PlainTextToDeviceEvent);
+            const toDeviceEvent = JSON.parse((processed as PlainTextToDeviceEvent).rawEvent);
 
             expect(toDeviceEvent.sender).toEqual("@alice:example.com");
             expect(toDeviceEvent.type).toEqual("custom.type");
@@ -1649,7 +1659,7 @@ describe(OlmMachine.name, () => {
                 content: {
                     "algorithm": "m.olm.v1.curve25519-aes-sha2",
                     "ciphertext": {
-                        aliceCurve: {
+                        [aliceCurve]: {
                             // this payload is just captured from a sync of some other element web with other users
                             body: "Awogjvpx458CGhuo77HX/+tp1sxgRoCi7iAlzMvfrpbWoREQAiKACysX/p+ojr5QitCi9WRXNyamW2v2LTvoyWKtVaA2oHnYGR5s5RYhDfnIgh5MMSqqKlAbfqLvrbLovTYcKagCBbFnbA43f6zYM44buGgy8q70hMVH5WP6aK1E9Z3DVZ+8PnXQGpsrxvz2IsL6w0Nzl/qUyBEQFcgkjoDPawbsZRCllMgq2LQUyqlun6IgDTCozqsfxhDWpdfYGde4z16m34Ang7f5pH+BmPrFs6E1AO5+UbhhhS6NwWlfEtA6/9yfMxWLz1d2OrLh+QG7lYFAU9/CzIoPxaHKKr4JxgL9CjsmUPyDymWOWHP0jLi1NwpOv6hGpx0FgM7jJIMk6gWGgC5rEgEeTIwdrJh3F9OKTNSva5hvD9LomGk6tZgzQG6oap1e3wiOUyTt6S7BlyMppIu3RlIiNihZ9e17JEGiGDXOXzMJ6ISAgvGVgTP7+EvyEt2Wt4du7uBo/UvljRvVNu3I8tfItizPAOlvz460+aBDxk+sflJWt7OnhiyPnOCfopb+1RzqKVCnnPyVaP2f4BPf8qpn/f5YZk+5jJgBrGPiHzzmb3sQ5pC470s6+U3MpVFlFTG/xPBtMRMwPsbKoHfnRPqIqGu5dQ1Sw7T6taDXWjP450TvjxgHK5t2z1rLA2SXzAB1P8xbi6YXqQwxL6PvMNHn/TM0jiIQHYuqg5/RKLyhHybfP8JAjgNBw9z16wfKR/YoYFr7c+S4McQaMNa8v2SxGzhpCC3duAoK2qCWLEkYRO5cMCsGm/9bf8Q+//OykygBU/hdkT1eHUbexgALPLdfhzduutU7pbChg4T7SH7euh/3NLmS/SQvkmPfm3ckbh/Vlcj9CsXws/7MX/VJbhpbyzgBNtMnbG6tAeAofMa6Go/yMgiNBZIhLpAm31iUbUhaGm2IIlF/lsmSYEiBPoSVfFU44tetX2I/PBDGiBlzyU+yC2TOEBwMGxBE3WHbIe5/7sKW8xJF9t+HBfxIyW1QRtY3EKdEcuVOTyMxYzq3L5OKOOtPDHObYiiXg00mAgdQqgfkEAIfoRCOa2NYfTedwwo0S77eQ1sPvW5Hhf+Cm+bLibkWzaYHEZF+vyE9/Tn0tZGtH07RXfUyhp1vtTH49OBZHGkb/r+L8OjYJTST1dDCGqeGXO3uwYjoWHXtezLVHYgL+UOwcLJfMF5s9DQiqcfYXzp2kEWGsaetBFXcUWqq4RMHqlr6QfbxyuYLlQzc/AYA/MrT3J6nDpNLcvozH3RcIs8NcKcjdtjvgL0QGThy3RcecJQEDx3STrkkePL3dlyFCtVsmtQ0vjBBCxUgdySfxiobGGnpezZYi7q+Xz61GOZ9QqYmkcZOPzfNWeqtmzB7gqlH1gkFsK2yMAzKq2XCDFHvA7YAT3yMGiY06FcQ+2jyg7Bk2Q+AvjTG8hlPlmt6BZfW5cz1qx1apQn1qHXHrgfWcI52rApYQlNPOU1Uc8kZ8Ee6XUhhXBGY1rvZiKjKFG0PPuS8xo4/P7/u+gH5gItmEVDFL6giYPFsPpqAQkUN7hFoGiVZEjO4PwrLOmydsEcNOfACqrnUs08FQtvPg0sjHnxh6nh6FUQv93ukKl6+c9d+pCsN2xukrQ7Dog3nrjFZ6PrS5J0k9rDAOwTB55sfGXPZ2rATOK1WS4XcpsCtqwnYm4sGNc8ALMQkQ97zCnw8TcQwLvdUMlfbqQ5ykDQpQD68fITEDDHmBAeTCjpC713E6AhvOMwTJvjhd7hSkeOTRTmn9zXIVGNo1jSr8u0xO9uLGeWsV0+UlRLgp7/nsgfermjwNN8wj6MW3DHGS8UzzYfe9TGCeywqqIUTqgfXY48leGgB7twh4cl4jcOQniLATTvigIAQIvq/Uv8L45BGnkpKTdQ5F73gehXdVA",
                             type: 1,
@@ -1664,22 +1674,79 @@ describe(OlmMachine.name, () => {
 
             let toDeviceEvents = [utdEvent];
 
-            const received = await m.receiveSyncChanges(
+            const receivedToDeviceArray = await m.receiveSyncChanges(
+                JSON.stringify(toDeviceEvents),
+                new DeviceLists(),
+                new Map<string, number>(),
+                undefined,
+            );
+            expect(receivedToDeviceArray.length).toBe(1);
+            const processed = receivedToDeviceArray[0];
+            expect(processed.type).toEqual(ProcessedToDeviceEventType.UnableToDecrypt);
+            expect(processed).toBeInstanceOf(UTDToDeviceEvent);
+            const toDeviceEvent = JSON.parse((processed as UTDToDeviceEvent).wireEvent);
+
+            expect(toDeviceEvent.sender).toEqual("@bob:example.org");
+            expect(toDeviceEvent.type).toEqual("m.room.encrypted");
+            expect(toDeviceEvent.content.ciphertext).toBeDefined();
+        });
+
+        test("Should return invalid event properly", async () => {
+            const m = await machine();
+
+            let eventWithNoType = {
+                sender: "@alice:example.com",
+                content: {
+                    algorithm: "m.megolm.v1.aes-sha2",
+                    code: "m.unverified",
+                    reason: "Device not verified",
+                    room_id: "!Cuyf34gef24t:localhost",
+                    sender_key: "RF3s+E7RkTQTGF2d8Deol0FkQvgII2aJDf3/Jp5mxVU",
+                    session_id: "X3lUlvLELLYxeTx4yOVu6UDpasGEVO0Jbu+QFnm0cKQ",
+                },
+            };
+
+            const eventWithMalformedCurveKey = {
+                content: {
+                    "algorithm": "m.olm.v1.curve25519-aes-sha2",
+                    "ciphertext": {
+                        INVALID_CURVE: {
+                            // this payload is just captured from a sync of some other element web with other users
+                            body: "Awogjvpx458CGhuo77HX/",
+                            type: 1,
+                        },
+                    },
+                    "org.matrix.msgid": "93ee0170aa7740d0ac9ee137e820302d",
+                    "sender_key": "WJ6Ce7U67a6jqkHYHd8o0+5H4bqdi9hInZdk0+swuXs",
+                },
+                type: "m.room.encrypted",
+                sender: "@bob:example.org",
+            };
+
+            let toDeviceEvents = [eventWithNoType, eventWithMalformedCurveKey];
+
+            const receivedToDeviceArray = await m.receiveSyncChanges(
                 JSON.stringify(toDeviceEvents),
                 new DeviceLists(),
                 new Map<string, number>(),
                 undefined,
             );
 
-            const receivedToDevices = JSON.parse(received);
-            expect(receivedToDevices).toBeInstanceOf(Array);
-            const receivedToDeviceArray: Array<any> = receivedToDevices;
-            expect(receivedToDeviceArray.length).toBe(1);
-            const toDeviceEvent = receivedToDeviceArray[0];
+            expect(receivedToDeviceArray.length).toBe(2);
 
-            expect(toDeviceEvent.sender).toEqual("@bob:example.org");
-            expect(toDeviceEvent.type).toEqual("m.room.encrypted");
-            expect(toDeviceEvent.content.ciphertext).toBeDefined();
+            const processed0 = receivedToDeviceArray[0];
+            expect(processed0.type).toEqual(ProcessedToDeviceEventType.Invalid);
+            expect(processed0).toBeInstanceOf(InvalidToDeviceEvent);
+            const toDeviceEvent0 = JSON.parse((processed0 as InvalidToDeviceEvent).wireEvent);
+            expect(toDeviceEvent0.sender).toEqual("@alice:example.com");
+            expect(toDeviceEvent0.content).toBeDefined();
+            expect(toDeviceEvent0.type).toBeUndefined();
+
+            const processed1 = receivedToDeviceArray[1];
+            expect(processed1.type).toEqual(ProcessedToDeviceEventType.Invalid);
+            const toDeviceEvent1 = JSON.parse((processed1 as InvalidToDeviceEvent).wireEvent);
+            expect(toDeviceEvent1.sender).toEqual("@bob:example.org");
+            expect(toDeviceEvent1.type).toEqual("m.room.encrypted");
         });
 
         test("Should return the clear text version of decrypted events", async () => {
@@ -1769,18 +1836,17 @@ describe(OlmMachine.name, () => {
 
             // Bob should be able to decrypt this
 
-            const received = await bob.receiveSyncChanges(
+            const receivedToDeviceArray = await bob.receiveSyncChanges(
                 JSON.stringify([encryptedToDevice]),
                 new DeviceLists(),
                 new Map<string, number>(),
                 undefined,
             );
-
-            const receivedToDevices = JSON.parse(received);
-            expect(receivedToDevices).toBeInstanceOf(Array);
-            const receivedToDeviceArray: Array<any> = receivedToDevices;
             expect(receivedToDeviceArray.length).toBe(1);
-            const toDeviceEvent = receivedToDeviceArray[0];
+            const processed = receivedToDeviceArray[0];
+            expect(processed.type).toEqual(ProcessedToDeviceEventType.Decrypted);
+            expect(processed).toBeInstanceOf(DecryptedToDeviceEvent);
+            const toDeviceEvent = JSON.parse((processed as DecryptedToDeviceEvent).decryptedRawEvent);
 
             expect(toDeviceEvent.sender).toEqual("@alice:example.org");
             expect(toDeviceEvent.type).toEqual("custom.type");